pyproject.toml
[project]
name = "smol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.15.2",
    "apscheduler==3.11.0",
    "fastapi==0.115.12",
    "ffmpeg-python==0.2.0",
    "hdbscan==0.8.40",
    "imagehash==4.3.2",
    "onnxruntime==1.21.1",
    "open-clip-torch==2.32.0",
    "opencv-python==4.11.0.86",
    "insightface>=0.7.3",
    "cmake>=3.27",
    "python-dateutil>=2.9.0.post0",
    "piexif==1.1.3",
    "pydantic==2.11.3",
    "pyinstaller>=6.15.0",
    "pyside6>=6.9.1",
    "python-multipart==0.0.20",
    "pywebview[qt]>=6.0",
    "qtpy>=2.4.1",
    "scenedetect==0.6.6",
    "sqlalchemy-vectorstores==0.1.3.post1",
    "sqlite-vec==0.1.6",
    "sqlmodel==0.0.24",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers==4.51.3",
    "uvicorn==0.34.2",
    "watchdog==6.0.0",
    "python-dateutil>=2.9.0.post0",
]
[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
torchaudio = [{ index = "pytorch-cpu" }]

# Declare the CPU index once
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
